# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FFgif API
  description: High-performance video-to-GIF conversion and sharing service built with Go, PostgreSQL, MinIO, Redis, RabbitMQ, and FFmpeg.
  version: 1.0.0
servers:
  - url: http://api:8080
    description: Docker network (used by ZAP scan)
  - url: http://localhost:8080
    description: Local dev
paths:
...
  /gifs/me:
    get:
      summary: List authenticated user's converted GIFs
      tags:
        - Media
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            default: all
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gif'
...
  /gifs/me/shares:
    get:
      summary: List GIFs shared with or by the authenticated user
      tags:
        - Share
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShareGifResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Gif:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        name:
          type: string
        user_id:
          type: string
          format: uuid
        url:
          type: string
        thumbnail_url:
          type: string
        file_size_bytes:
          type: integer
        width:
          type: integer
        height:
          type: integer
        duration_seconds:
          type: number
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ShareGifResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
        thumbnail_url:
          type: string
        gif_key:
          type: string
        owner_id:
          type: string
          format: uuid
        shared_with:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
...